Segregating groups of tests into separate files
