+++ conin/dynamic_bayesian_network/tests/test_weather.py
@@ -1,5 +1,5 @@
 import pytest
-from . import test_cases
+from . import examples
 
 try:
     import pgmpy
@@ -15,170 +15,11 @@ from conin.dynamic_bayesian_network import (
 
 
 @pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
-def test_simple0_ALL():
-    """
-    Z
-    """
-    G = test_cases.simple0_DBN()
-    q = {("Z", 0): 1, ("Z", 1): 0}
-
-    model = create_DBN_map_query_model(pgm=G)  # variables=None, evidence=None
-    results = optimize_map_query_model(model, solver="glpk")
-    assert q == results.solution.variable_value
-
-
-@pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
-def test_simple2_ALL():
-    """
-    Z
-    """
-    G = test_cases.simple2_DBN()
-    q = {("Z", 0): 1, ("Z", 1): 0}
-
-    model = create_DBN_map_query_model(pgm=G)  # variables=None, evidence=None
-    results = optimize_map_query_model(model, solver="glpk")
-    assert q == results.solution.variable_value
-
-
-@pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
-def test_simple3_ALL():
-    """
-    Z
-    """
-    G = test_cases.simple3_DBN()
-    q = {("Z", 0): 1, ("Z", 1): 0}
-
-    model = create_DBN_map_query_model(pgm=G)  # variables=None, evidence=None
-    results = optimize_map_query_model(model, solver="glpk")
-    assert q == results.solution.variable_value
-
-
-@pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
-def test_simple1_ALL():
-    """
-    A -> B
-
-    No evidence
-    """
-    G = test_cases.simple1_DBN()
-    q = {
-        ("A", 0): 0,
-        ("A", 1): 1,
-        ("B", 0): 1,
-        ("B", 1): 0,
-    }
-
-    model = create_DBN_map_query_model(pgm=G)  # variables=None, evidence=None
-    results = optimize_map_query_model(model, solver="glpk")
-    assert q == results.solution.variable_value
-
-
-@pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
-def test_simple1_B():
-    """
-    A -> B
-
-    Evidence: A_0 = 1
-    """
-    G = test_cases.simple1_DBN()
-    q = {
-        ("A", 1): 0,
-        ("B", 0): 0,
-        ("B", 1): 1,
-    }
-
-    model = create_DBN_map_query_model(
-        pgm=G, evidence={("A", 0): 1}
-    )  # variables=None, evidence=None
-    results = optimize_map_query_model(model, solver="glpk")
-    assert q == results.solution.variable_value
-
-
-@pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
-def test_simple1_ALL_constrained():
-    """
-    A -> B
-
-    No evidence
-    """
-    cpgm = test_cases.simple1_DBN_constrained()
-    q = {
-        ("A", 0): 0,
-        ("A", 1): 0,
-        ("B", 0): 1,
-        ("B", 1): 1,
-    }
-
-    results = optimize_map_query_model(cpgm.create_map_query_model(), solver="glpk")
-    assert q == results.solution.variable_value
-
-
-@pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
-def test_simple3_ALL():
-    """
-    A -> B
-
-    No evidence
-    """
-    G = test_cases.simple3_DBN()
-    q = {
-        ("A", 0): 0,
-        ("A", 1): 1,
-        ("B", 0): 1,
-        ("B", 1): 0,
-    }
-
-    model = create_DBN_map_query_model(pgm=G)  # variables=None, evidence=None
-    results = optimize_map_query_model(model, solver="glpk")
-    assert q == results.solution.variable_value
-
-
-@pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
-def test_simple3_B():
-    """
-    A -> B
-
-    Evidence: A_0 = 1
-    """
-    G = test_cases.simple3_DBN()
-    q = {
-        ("A", 1): 0,
-        ("B", 0): 0,
-        ("B", 1): 1,
-    }
-
-    model = create_DBN_map_query_model(
-        pgm=G, evidence={("A", 0): 1}
-    )  # variables=None, evidence=None
-    results = optimize_map_query_model(model, solver="glpk")
-    assert q == results.solution.variable_value
-
-
-@pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
-def test_simple3_ALL_constrained():
-    """
-    A -> B
-
-    No evidence
-    """
-    cpgm = test_cases.simple3_DBN_constrained()
-    q = {
-        ("A", 0): 0,
-        ("A", 1): 0,
-        ("B", 0): 1,
-        ("B", 1): 1,
-    }
-
-    results = optimize_map_query_model(cpgm.create_map_query_model(), solver="glpk")
-    assert q == results.solution.variable_value
-
-
-@pytest.mark.skipif(not pgmpy_available, reason="pgmpy not installed")
 def test_weather1_A():
     """
     Test with TabularCPD representation
     """
-    pgm = test_cases.pgmpy_weather1()
+    pgm = examples.pgmpy_weather1()
     q = {
         ("H", 0): "Low",
         ("H", 1): "Low",
@@ -212,7 +53,7 @@ def test_weather1_B():
     """
     Test with TabularCPD representation
     """
-    pgm = test_cases.pgmpy_weather1()
+    pgm = examples.pgmpy_weather1()
     evidence = {
         ("O", 0): "Wet",
         ("O", 1): "Wet",
@@ -248,7 +89,7 @@ def test_weather2_A():
     """
     Test with TabularCPD representation
     """
-    pgm = test_cases.pgmpy_weather2()
+    pgm = examples.pgmpy_weather2()
     q = {
         ("H", 0): "Low",
         ("H", 1): "Low",
@@ -282,7 +123,7 @@ def test_weather2_B():
     """
     Test with TabularCPD representation
     """
-    pgm = test_cases.pgmpy_weather2()
+    pgm = examples.pgmpy_weather2()
     evidence = {
         ("O", 0): "Wet",
         ("O", 1): "Wet",
@@ -318,7 +159,7 @@ def test_weather2_B_constrained1():
     """
     Test with TabularCPD representation
     """
-    cpgm = test_cases.pgmpy_weather_constrained1()
+    cpgm = examples.pgmpy_weather_constrained1()
     evidence = {
         ("O", 0): "Wet",
         ("O", 1): "Wet",
