+++ conin/markov_network/tests/examples.py
