+++ conin/markov_network/tests/test_abc.py
@@ -11,7 +11,7 @@ from conin.markov_network.factor_repn import (
 from conin.markov_network.inference import (
     create_MN_map_query_model_from_factorial_repn,
 )
-from . import test_cases
+from . import examples
 
 try:
     import pgmpy
@@ -19,59 +19,6 @@ try:
     pgmpy_available = True
 except Exception as e:
     pgmpy_available = False
-
-
-def test_example6():
-    """
-    See Obbens, p.18
-
-    Non-uniform weights are applied to the A_B factors to remove a degeneracy w.r.t. the
-    value of variable A.
-    """
-    S = {"A": [State(0), State(1)], "B": [State(0), State(1)]}
-
-    J = {"A": [0, 1], "B": [0, 1], "A_B": [0, 1, 2, 3]}
-
-    v = {
-        ("A", 0, "A"): State(0),
-        ("A", 1, "A"): State(1),
-        ("B", 0, "B"): State(0),
-        ("B", 1, "B"): State(1),
-        ("A_B", 0, "A"): State(0),
-        ("A_B", 0, "B"): State(0),
-        ("A_B", 1, "A"): State(0),
-        ("A_B", 1, "B"): State(1),
-        ("A_B", 2, "A"): State(1),
-        ("A_B", 2, "B"): State(0),
-        ("A_B", 3, "A"): State(1),
-        ("A_B", 3, "B"): State(1),
-    }
-
-    w = {
-        ("A", 0): -log(2),
-        ("A", 1): -log(2),
-        ("B", 0): -log(3),
-        ("B", 1): log(2 / 3),
-        ("A_B", 0): -log(6),
-        ("A_B", 1): -log(2),
-        ("A_B", 2): -log(6),
-        ("A_B", 3): -log(6),
-    }
-
-    model = create_MN_map_query_model_from_factorial_repn(S=S, J=J, v=v, w=w)
-    results = optimize_map_query_model(model, solver="glpk")
-    assert results.solution.variable_value == {"A": 0, "B": 1}
-
-    if pgmpy_available:
-        pgm = test_cases.example6()
-        S_, J_, v_, w_ = extract_factor_representation(pgm)
-        assert S == S_
-        assert J == J_
-        assert v == v_
-        assert w == w_
-        model = create_MN_map_query_model(pgm=pgm)
-        results = optimize_map_query_model(model, solver="glpk")
-        assert results.solution.variable_value == {"A": 0, "B": 1}
 
 
 def test_ABC():
@@ -208,7 +155,7 @@ def test_ABC():
     assert results.solution.variable_value == {"A": 2, "B": 2, "C": 1}
 
     if pgmpy_available:
-        pgm = test_cases.ABC()
+        pgm = examples.ABC()
         S_, J_, v_, w_ = extract_factor_representation(pgm)
         assert S == S_
         assert J == J_
@@ -363,7 +310,7 @@ def test_ABC_constrained():
 
     if pgmpy_available:
         # Double-check that we can extract the right factor representation
-        cpgm = test_cases.ABC_constrained()
+        cpgm = examples.ABC_constrained()
         S_, J_, v_, w_ = extract_factor_representation(cpgm.pgm)
         assert S == S_
         assert J == J_
@@ -372,7 +319,7 @@ def test_ABC_constrained():
 
         # Constrain the inference to ensure that all variables have different
         # values
-        pgm = test_cases.ABC()
+        pgm = examples.ABC()
         model = create_MN_map_query_model(pgm=pgm)
 
         def diff_(M, s):
@@ -386,6 +333,6 @@ def test_ABC_constrained():
 
         # Constrain the inference to ensure that all variables have different
         # values
-        cpgm = test_cases.ABC_constrained()
+        cpgm = examples.ABC_constrained()
         results = optimize_map_query_model(cpgm.create_map_query_model(), solver="glpk")
         assert results.solution.variable_value == {"A": 0, "B": 2, "C": 1}
